sidecar: add proxy-port for including alternative public accessible port
The new `proxy-port` argument can be used to specify an alternative port
to the `controller-port` where public connections are accepted. This
port is included in the CSIAddonsNode.Endpoint which is used by the
Kubernetes CSI-Addons controller to connect to the sidecar.

The proxy-port will be used when connections are routed through a proxy
for encrypting the gRPC communication and extra authentication checks.

Signed-off-by: Niels de Vos <ndevos@ibm.com>